SQL_Problem_1
Puzzle #1
Dance Partners
You are tasked with providing a list of dance partners from the following table.
Provide an SQL statement that matches each Student ID with an individual of the opposite gender.
Note there is a mismatch in the number of students, as one female student will be left without a dance partner. Please include this individual in your list as well.